add validator and specific errors handler

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { Idea } from './idea.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -23,16 +23,31 @@ export class IdeaService {
     }
 
     async read(id: number) {
-        return await this.ideaRepository.findOne({ where : { id }});
+        const idea = await this.ideaRepository.findOne({ where : { id }});
+        console.log(idea);
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
+        return idea;
     }
 
     async update(id: number, data: Partial<IdeaDTO>) {
+        const idea = await this.ideaRepository.findOne({ where : { id }});
+        console.log(idea);
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.update({ id }, data);
         return await this.ideaRepository.findOne({ where : { id }});
     }
 
     async destroy(id: number) {
+        const idea = await this.ideaRepository.findOne({ where : { id }});
+        console.log(idea);
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.delete({ id });
-        return { deleted : true };
+        return idea;
     }
 }
